extract container repairing logic to orient db module
